create new module telegram-bot-api and move files
