+++ telegram-bot-api/src/main/kotlin/com/ajaxproject/telegrambot/bot/dto/MonobankCurrencyExchangeResponse.kt
