+++ telegram-bot-api/src/main/kotlin/com/ajaxproject/telegrambot/bot/enums/ConversationState.kt
@@ -4,6 +4,5 @@ enum class ConversationState {
 
     CONVERSATION_STARTED,
     WAITING_FOR_NUMBER,
-    WAITING_FOR_ADD_INCOME,
-    WAITING_FOR_ADD_EXPENSE
+    WAITING_FOR_ADD_FINANCE,
 }
