+++ telegram-bot-api/src/main/kotlin/com/ajaxproject/telegrambot/bot/enums/Currency.kt
