+++ telegram-bot-api/src/main/kotlin/com/ajaxproject/telegrambot/bot/enums/TextPropertyName.kt
@@ -7,7 +7,7 @@ enum class TextPropertyName {
     MENU_TEXT,
     CURRENCY_TEXT,
     FINANCE_MANAGE_TEXT,
-    ADD_FINANCE,
+    ADD_FINANCE_TEXT,
     SUCCESSFUL_ADD_FINANCE,
     FAILED_ADD_FINANCE,
     BACK_TO_MENU
