+++ telegram-bot-api/src/main/kotlin/com/ajaxproject/telegrambot/bot/handlers/impl/GetCurrencyExchangeRateHandler.kt
