+++ telegram-bot-api/src/main/kotlin/com/ajaxproject/telegrambot/bot/models/MongoCurrency.kt
