+++ telegram-bot-api/src/main/kotlin/com/ajaxproject/telegrambot/bot/properties/ApplicationProperties.kt
