+++ telegram-bot-api/src/main/kotlin/com/ajaxproject/telegrambot/bot/service/updatemodels/UpdateSession.kt
